Add a flag that allows to accept unknown audiences in the ID token.
Add a flag to the list of supported environment variables (ACCEPT_UNKNOWN_AUDIENCES)
that, when set to true, will ignore unknown audiences in the ID token.
The list of audiences will still be checked for the audience set in CLIENT_ID.

This enables the use of IDPs that set more than one audience in the ID token
(e.g., Zitadel).