Add a flag that allows to provide a list of accepted audiences.
Add a flag that can be set to provide a list of accepted audiences,
beyond its own audience (i.e., from CLIENT_ID). If this is set, it
must contain all the additional audiences the IDP might send.

This enables the use of IDPs that set more than one audience in the
ID token (e.g., Zitadel). However, specifically for Zitadel, the flag
needs to be updated whenever a new application is added to the same
project. If this happens often, using ACCEPT_UNKNOWN_AUDIENCES may
be the only choice.